Tests/LibWeb/Text/expected/css/effective-color-scheme-resolution.txt

preferred: light dark / rgb(235, 235, 235)
first-supported: custom light / rgb(0, 0, 0)
normal: normal / rgb(235, 235, 235)
variable: dark / rgb(235, 235, 235)
animated: dark / rgb(0, 0, 0)
